Inlined props for anchor elements; inline elements with more props than children can use this API so they can be read from the DOM
The reason we are inlining props when needed is because the editor is fundamentally a controlled and uncontrolled element at the same time. Notably, the input handler is uncontrolled, and we need to reactively read back a UUID element’s spans after the input event is fired. This affords us the ability to not interfere with user input most of the time (sans backspace, etc. -- more controlled events). The drawback with this approach is that we need to be able to rebuild the React VDOM from the DOM. This could be problematic for invisible props unless we literally encode them into elements.

In Chrome, when you type inside of an anchor element in a contenteditable context, the surrounding <a> tag is destroyed (in our case, just leaving the inside span). This is why we can’t simply query href from DOM, because it literally doesn’t exist anymore. This is why inlining props (via JSON.stringify) on the remaining <span> element that persists is a reasonable solution, albeit weird if someone were to inspect the DOM.

It’s worth mentioning that most editors seem to not do this because their inline elements are more simple in scope -- styles (e.g. font-style: italic and font-weight: 700) suffice in place of React components. Because we want to potentially add support for custom inline elements, we want to get into the practice of authoring components as opposed to faking it with styles alone.