Documenting a bug
Because contenteditable-edited anchor elements remove the parent <a> element, React gets confused and leads to duplicating input. This can potentially be avoided by using JavaScript to create a synthetic anchor.

Note that Notion does not do this -- Notion actually does not extend an anchor when you type at the right-side (e.g. end) of a link. This behavior is inconsistent with other inline elements, so it seems it should be avoided despite being more semantically correct (this is essentially the same argument for why to use <span> elements instead of semantic elements).