Documenting a bug
Note that flex and inline-flex classes disrupt the arrow keys from working as expected; arrow keys jump to the wrong positions. This can be either handled with synthetic handlers (preferable) or more careful CSS. Note that more careful CSS does not suggest this fixes the problem indefinitely, because this is more or less up to the host browser.

If we do add synthetic handlers for arrow keys, we likely only need to implement them for word or at most word boundaries.