More renaming
The convention is basically to give files with multiple exports a general name so they easier to find, and files that only ever export default the name of the export so it’s clear when importing what the desirable module name is.

Generally, things should be plural unless it doesn’t make sense or if the word in it of itself is already plural in nature or could be construed as plural e.g. "whitespace". The reason for this is "whitespaces" is more awkward than "whitespace" alone.