Fixed most compatibility bugs
When using whitespace-pre-wrap (class or style) on the <article> element, Firefox selects / focuses absolute elements when arrowing-up or arrowing-down out of block elements. By moving whitespace-pre-wrap to the <Element> HOC component, this is no longer the case. Note that absolute elements should use contentEditable={false}.

Note that it was previously observed that pre-wrap must be applied as a style and **does not** work as a class. I don’t know if this is still the case, but it seems to be more self-documenting to use inline styles given these are imperative styles that *must* be applied. See https://github.com/codex-src/codex-web-app/commit/56f54d6e923a361796931778c4db56c0283f2a00 for previous reference.

Furthermore, whitespace-pre-wrap is needed in general because (observed in Firefox) browsers can use &nbsp; (code point 160) in place of common spaces (code point 32), which is undesirable.

Finally, break-words (observed in Firefox) breaks up non-breaking words that overrun a given line. Note that this was **not** observed in Chrome.

Therefore, applying white-space: pre-wrap and overflow-wrap: break-word as inline styles on <Element> HOC component is needed for a myriad of reasons.