I think the rendering issue stems from ref = toArray(ref).slice(-1)[0]
Basically, we can render an array with intermediary elements, but we can also render an array with strings, which means if we do ref = toArray(ref).slice(-1)[0], there’s a very real probability we land on a string. We already have a guard in place to eagerly return for strings, but commenting out this line of code **does** render the missing text.