Better implementation of display markdown mode; this version relies on the presence of immutable DOM elements; markdown elements are set to contenteditable=false and the parsers have been updated not to interact with such elements (see test suites). Furthermore, this version makes copying the markdown text automatic as a side-effect.
Technically, this version renders a more complicated DOM but is simpler from the perspective of side-effects. The ::before and ::after version has a simpler DOM implementation but does not render selection backgrounds and does not implicitly solve copying text as markdown. It seems like the React-rendered implementation is slightly more advantageous, despite that it emits a more complicated DOM.

This would **not** be the case if we could simply use JavaScript to conditionally render markdown elements.