The aforementioned idea doesn’t really work because components are not transparent like that, at least memoized components aren’t. So alternatively, maybe we can just analyze the rendered DOM or simply call toReactTree twice.
For now, let’s just call toReactTree twice. This isn’t a terrible solution since we’re treating the console as a separate app from the editor itself. So the editor technically isn’t supported to know it’s rendered state. Given that we’re already debouncing elements this should be OK for now.