Prevented GFM code elements from recursing; returns treeTextContent now
Note that right now we don’t need to worry about **`code`** cases because PrismJS does not provide enough context for our CSS to style correctly. Idealistically we want to emit a markdown snippet ourselves and style against our generated classes. We also want to do this for HTML. In general, we get better flexibility / control over *how* we style elements going forward *and* parsing should be much faster if we’re careful, which means we can reduce the debouncer. This contributes to a smoother UX.